move (most of the) implementations depending on native code to native/
