+++ objectbox/lib/src/native/bindings/data_visitor.dart
