+++ objectbox/lib/src/native/bindings/flatbuffers.dart
