+++ objectbox/lib/src/native/bindings/helpers.dart
