+++ objectbox/lib/src/native/bindings/structs.dart
