+++ objectbox/lib/src/native/model.dart
