+++ objectbox/lib/src/native/query/property.dart
