+++ objectbox/lib/src/native/store.dart
