+++ objectbox/lib/src/native/transaction.dart
